change coordinate types to doubles, display vehicles as rectangles, add some tests

+++ brstu-traffic-simulation-tool/core/src/main/java/by/brstu/tst/core/map/primitives/MapPoint.java
@@ -6,25 +6,25 @@ package by.brstu.tst.core.map.primitives;
  * Coordinates in EPSG:3785
  */
 public class MapPoint {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public MapPoint(float lon, float lat) {
+    public MapPoint(double lon, double lat) {
         this.x = lon;
         this.y = lat;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public float calculateDistance(MapPoint point) {
-        float dlon = x - point.x;
-        float dlat = y - point.y;
-        return (float) Math.sqrt(dlon * dlon + dlat * dlat);
+    public double calculateDistance(MapPoint point) {
+        double dlon = x - point.x;
+        double dlat = y - point.y;
+        return Math.sqrt(dlon * dlon + dlat * dlat);
     }
 }
